deep_agent/pyproject.toml
[project]
name = "deep-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "deepagents>=0.4.4",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
]
